requirements.txt
python-telegram-bot>=21.0
faster-whisper>=1.1.1,<2
# Optional Remotion renderer dependencies are installed separately with: cd remotion && npm install
